# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Salud API",
        "description": "Documentación de la API de Gestión de Salud",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/movimientos-farmacia": {
            "get": {
                "tags": [
                    "Farmacia"
                ],
                "summary": "Listar movimientos de farmacia",
                "operationId": "af2cc3fe9220493e54e87ea82e49f840",
                "parameters": [
                    {
                        "name": "medicamento_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "receta_medica_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "farmaceutico_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tipo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "entrada",
                                "salida"
                            ]
                        }
                    },
                    {
                        "name": "fecha_desde",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_hasta",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de movimientos"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Farmacia"
                ],
                "summary": "Registrar movimiento de farmacia (entrada o salida)",
                "operationId": "4ed3362e27115b5854c727a205f7c8f4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "medicamento_id",
                                    "tipo",
                                    "cantidad",
                                    "fecha"
                                ],
                                "properties": {
                                    "medicamento_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "receta_medica_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "farmaceutico_id": {
                                        "type": "integer",
                                        "example": 5,
                                        "nullable": true
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "entrada",
                                        "enum": [
                                            "entrada",
                                            "salida"
                                        ]
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "detalle": {
                                        "type": "string",
                                        "example": "Compra mensual",
                                        "nullable": true
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movimiento registrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "409": {
                        "description": "Stock insuficiente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/movimientos-farmacia/{id}": {
            "get": {
                "tags": [
                    "Farmacia"
                ],
                "summary": "Detalle de movimiento de farmacia",
                "operationId": "d6b93b5fcfb3ec57277e65c254bc08d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movimiento encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/medicamentos/{medicamento_id}/movimientos": {
            "get": {
                "tags": [
                    "Farmacia"
                ],
                "summary": "Movimientos de un medicamento específico",
                "operationId": "9ea235b7b9478c07355b5f5c9394bf88",
                "parameters": [
                    {
                        "name": "medicamento_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movimientos del medicamento"
                    },
                    "404": {
                        "description": "Medicamento no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Usa el formato Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}